Add a few graphql bindings for addresses
