+++ lib/cambiatus_web/schema/kyc_types.ex
@@ -1,4 +1,4 @@
-defmodule CambiatusWeb.Schema.KycType do
+defmodule CambiatusWeb.Schema.KycTypes do
   @moduledoc """
   This module hold GraphQL objects related to the KYC process
   """
@@ -6,13 +6,14 @@ defmodule CambiatusWeb.Schema.KycType do
   use Absinthe.Schema.Notation
   use Absinthe.Relay.Schema.Notation, :classic
   import Absinthe.Resolution.Helpers, only: [dataloader: 1]
+  alias CambiatusWeb.Resolvers.Kyc
 
   @desc "Address query"
   object :address_queries do
     @desc "List of supported countries"
-    field(:countries, non_null(list_of(non_null(:country)))) do
+    field(:country, :country) do
       arg(:input, non_null(:country_input))
-      # TODO: do resolver
+      resolve(&Kyc.get_country/3)
     end
   end
 
@@ -24,19 +25,19 @@ defmodule CambiatusWeb.Schema.KycType do
   @desc "KYC supported countries"
   object :country do
     field(:name, non_null(:string))
-    field(:states, non_null(list_of(:state)), resolve: dataloader(Cambiatus.Kyc))
+    field(:states, non_null(list_of(:state)), resolve: dataloader(Kyc))
   end
 
   @desc "KYC supported states"
   object :state do
     field(:name, non_null(:string))
-    field(:cities, non_null(list_of(:cities)), resolve: dataloader(Cambiatus.Kyc))
+    # field(:cities, non_null(list_of(:city)), resolve: dataloader(Cambiatus.Kyc))
   end
 
   @desc "KYC supported cities"
   object :city do
     field(:name, non_null(:string))
-    field(:neighborhoods, non_null(list_of(:neighborhood)), resolve: dataloader(Cambiatus.Kyc))
+    # field(:neighborhoods, non_null(list_of(:neighborhood)), resolve: dataloader(Cambiatus.Kyc))
   end
 
   @desc "KYC supported neighborhoods"
